Fix entrypoint.js windows elevate
